replaced add student with ng generated component
