+++ src/app/components/add-student/add-student.component.html
@@ -1,0 +1,10 @@
+<div class="create-student">
+    <h4>Create new Student</h4>
+    <label for="firstNameInput">First Name</label>
+    <input name="firstNameInput" type="text" [(ngModel)] = "firstName" >
+    <label for="lastNameInput">Last Name</label>
+    <input name="lastNameInput" type="text" [(ngModel)] = "lastName" >
+    <label for="guardianInput">Guardian Email Address</label>
+    <input name="guardianInput" type="text" [(ngModel)] = "guardianEmail" >
+    <button (click)="createStudent()">Submit</button>
+</div>
